兑吧积分订单表字段属性修改

+++ hyjf_am_server/hyjf_am_core/hyjf_am_market/src/main/java/com/hyjf/am/market/dao/model/auto/DuibaOrders.java
@@ -150,7 +150,7 @@ public class DuibaOrders implements Serializable {
      *
      * @mbggenerated
      */
-    private BigDecimal integralPrice;
+    private Integer integralPrice;
 
     /**
      * 创建人
@@ -363,11 +363,11 @@ public class DuibaOrders implements Serializable {
         this.exchangeRate = exchangeRate;
     }
 
-    public BigDecimal getIntegralPrice() {
+    public Integer getIntegralPrice() {
         return integralPrice;
     }
 
-    public void setIntegralPrice(BigDecimal integralPrice) {
+    public void setIntegralPrice(Integer integralPrice) {
         this.integralPrice = integralPrice;
     }
